Adjust to `git-discovery` to support linked worktrees in `Kind` type (#301)
